+++ lib/providers/match_madness_provider.g.dart
@@ -1,26 +1,27 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of 'mistake_provider.dart';
+part of 'match_madness_provider.dart';
 
 // **************************************************************************
 // RiverpodGenerator
 // **************************************************************************
 
-String _$mistakeNotifierHash() => r'b190fcb2b76596103a627e993ab6667ce6dc782c';
+String _$matchMadnessControllerHash() =>
+    r'fe14f338395bd56d4382697ad64dac08ee6f29cf';
 
-/// See also [MistakeNotifier].
-@ProviderFor(MistakeNotifier)
-final mistakeNotifierProvider = AutoDisposeAsyncNotifierProvider<
-    MistakeNotifier, List<SentencePair>>.internal(
-  MistakeNotifier.new,
-  name: r'mistakeNotifierProvider',
+/// See also [MatchMadnessController].
+@ProviderFor(MatchMadnessController)
+final matchMadnessControllerProvider = AutoDisposeNotifierProvider<
+    MatchMadnessController, MatchMadnessState>.internal(
+  MatchMadnessController.new,
+  name: r'matchMadnessControllerProvider',
   debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
       ? null
-      : _$mistakeNotifierHash,
+      : _$matchMadnessControllerHash,
   dependencies: null,
   allTransitiveDependencies: null,
 );
 
-typedef _$MistakeNotifier = AutoDisposeAsyncNotifier<List<SentencePair>>;
+typedef _$MatchMadnessController = AutoDisposeNotifier<MatchMadnessState>;
 // ignore_for_file: type=lint
 // ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member, deprecated_member_use_from_same_package