+++ lib/screens/screens.dart
@@ -3,3 +3,5 @@ export 'home_screen.dart';
 export 'lesson_screen.dart';
 export 'leaderboard_screen.dart';
 export 'profile_screen.dart';
+export 'match_madness_screen.dart';
+